Clean stale bloop files after import
After amending or refreshing a project, some targets may no longer
exist, and their Bloop configuration files should be removed. This
commit ports over the logic that was taking care of that in Fastpass for
Pants to Fastpass for Bazel.
